# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DamascusTransit API",
    "description": "Real-time transit tracking and fleet management — multi-tenant SaaS edition.\n\n## Authentication\n\nMost read endpoints are public but require an `?operator=<slug>` query parameter.\n1. **POST /api/v1/auth/login** — exchange email/password for a JWT token\n2. Include the token as `Authorization: Bearer <token>` on protected endpoints\n\nRoles: `super_admin` · `admin` · `dispatcher` · `driver` · `viewer`\n\nTokens expire after 24 hours.",
    "contact": {
      "name": "Damascus Transit System",
      "email": "support@damascustransit.com"
    },
    "license": {
      "name": "MIT"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/vehicles": {
      "get": {
        "tags": [
          "vehicles"
        ],
        "summary": "List Vehicles",
        "description": "List all active vehicles with latest positions.",
        "operationId": "list_vehicles_api_vehicles_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "operator",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Operator slug",
              "title": "Operator"
            },
            "description": "Operator slug"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VehicleResponse"
                  },
                  "title": "Response List Vehicles Api Vehicles Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/vehicles/positions": {
      "get": {
        "tags": [
          "vehicles"
        ],
        "summary": "Get Vehicle Positions",
        "description": "Get latest vehicle positions only (lightweight endpoint).",
        "operationId": "get_vehicle_positions_api_vehicles_positions_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "operator",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Operator slug",
              "title": "Operator"
            },
            "description": "Operator slug"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response Get Vehicle Positions Api Vehicles Positions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/traccar/position": {
      "post": {
        "tags": [
          "traccar"
        ],
        "summary": "Traccar Position Webhook",
        "description": "Webhook for Traccar GPS position updates. Secured by HMAC signature.",
        "operationId": "traccar_position_webhook_api_traccar_position_post",
        "parameters": [
          {
            "name": "x-traccar-signature",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "description": "HMAC-SHA256 signature",
              "title": "X-Traccar-Signature"
            },
            "description": "HMAC-SHA256 signature"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TraccarPosition"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/traccar/event": {
      "post": {
        "tags": [
          "traccar"
        ],
        "summary": "Traccar Event Webhook",
        "description": "Webhook for Traccar events (engine on/off, speeding, etc). Secured by HMAC.",
        "operationId": "traccar_event_webhook_api_traccar_event_post",
        "parameters": [
          {
            "name": "x-traccar-signature",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "description": "HMAC-SHA256 signature",
              "title": "X-Traccar-Signature"
            },
            "description": "HMAC-SHA256 signature"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TraccarEvent"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TraccarEvent": {
        "properties": {
          "eventId": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Eventid"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "serverTime": {
            "type": "integer",
            "title": "Servertime"
          },
          "deviceId": {
            "type": "integer",
            "title": "Deviceid"
          },
          "deviceName": {
            "type": "string",
            "title": "Devicename"
          },
          "data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "type",
          "serverTime",
          "deviceId",
          "deviceName",
          "data"
        ],
        "title": "TraccarEvent"
      },
      "TraccarPosition": {
        "properties": {
          "deviceId": {
            "type": "integer",
            "title": "Deviceid"
          },
          "latitude": {
            "type": "number",
            "title": "Latitude"
          },
          "longitude": {
            "type": "number",
            "title": "Longitude"
          },
          "altitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Altitude"
          },
          "speed": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Speed"
          },
          "heading": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Heading"
          },
          "accuracy": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accuracy"
          },
          "timestamp": {
            "type": "integer",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "deviceId",
          "latitude",
          "longitude",
          "timestamp"
        ],
        "title": "TraccarPosition"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "VehicleResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "vehicle_id": {
            "type": "string",
            "title": "Vehicle Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "name_ar": {
            "type": "string",
            "title": "Name Ar"
          },
          "vehicle_type": {
            "type": "string",
            "title": "Vehicle Type"
          },
          "capacity": {
            "type": "integer",
            "title": "Capacity"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "assigned_route_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned Route Id"
          },
          "assigned_driver_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned Driver Id"
          },
          "latitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude"
          },
          "longitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude"
          },
          "speed_kmh": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Speed Kmh"
          },
          "occupancy_pct": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Occupancy Pct"
          },
          "recorded_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recorded At"
          },
          "approval_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approval Status"
          },
          "approved_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approved At"
          },
          "approval_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approval Note"
          },
          "driver_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Driver Name"
          },
          "driver_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Driver Email"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "vehicle_id",
          "name",
          "name_ar",
          "vehicle_type",
          "capacity",
          "status"
        ],
        "title": "VehicleResponse"
      },
...
      "WebhookResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "timestamp": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timestamp"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "WebhookResponse"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}